feat(hello_world): let terraform manage lambda logging group
Normally, a lambda function would automatically create a new logging group in
cloudwatch. This log group will therefore not be managed by terraform. It will
not be deleted on removing the lambda function and only has default configuration
parameter (i.e. rentention time would be infinite).

The solution is to deny all lambda function to create a new log group
(`logs:CreateLogGroup`) and create the group with terraform itself.